+++ app/ExternalFiles/Forms.php
@@ -266,7 +266,7 @@ class Forms {
 				<p>
 					<?php
 					/* translators: %1$s will be replaced with the URL for add new media */
-					echo wp_kses_post( sprintf( __( 'Add external files via their URL <a href="%1$s" target="_blank">here (opens in a new window)</a>.', 'external-files-in-media-library' ), esc_url( $url ) ) );
+					echo wp_kses_post( sprintf( __( 'Add external files via their URL <a href="%1$s" target="_blank">here%2$s</a>.', 'external-files-in-media-library' ), esc_url( $url ), Helper::get_a11n_window_hint() ) );
 					?>
 				</p>
 			</div>